Revert "8259635: Update to 610.2 version of WebKit"
This reverts commit 9d17327ffa6a7b59df13017a48e6bd858836f30a.

+++ modules/web/src/main/native/Source/WebCore/inspector/NetworkResourcesData.h
@@ -99,7 +99,7 @@ public:
         bool hasData() const { return m_dataBuffer; }
         size_t dataLength() const;
         void appendData(const char* data, size_t dataLength);
-        unsigned decodeDataToContent();
+        size_t decodeDataToContent();
 
         String m_requestId;
         String m_loaderId;
